Support copying asset-directories, not just files (#16)
* add https://stackoverflow.com/a/26038979/4260616

* corrections

* corrections

* corrections

* Update build.js

+++ npm/scripts/build.js
@@ -81,8 +81,13 @@ function buildAssets(root, assets, includeVendor)
 			}
 
 			// On the first entry write the file
-			if (index == 0) {
+			if (index == 0 && fs.statSync(file).isFile()) {
 				fs.copyFileSync(file, root + '/' + asset.dest);
+				return;
+			}
+			
+			if (index == 0) {
+				copyFolderRecursiveSync(file, root + '/' + asset.dest);
 				return;
 			}
 
@@ -141,6 +146,43 @@ function buildAssets(root, assets, includeVendor)
 		if (asset.onlyMinified === true) {
 			fs.unlinkSync(root + '/' + asset.dest);
 		}
+	});
+}
+
+function copyFileSync(source, target)
+{
+	let targetFile = target;
+
+	// If target is a directory, a new file with the same name will be created
+	if (fs.existsSync(target) && fs.lstatSync(target).isDirectory()) {
+		targetFile = path.join(target, path.basename(source));
+	}
+
+	fs.writeFileSync(targetFile, fs.readFileSync(source));
+}
+
+function copyFolderRecursiveSync(source, target)
+{
+	// Check if folder needs to be created or integrated
+	const targetFolder = path.join(target, path.basename(source));
+	if (!fs.existsSync(targetFolder)) {
+		fs.mkdirSync(targetFolder);
+	}
+
+	// Copy
+	if (!fs.lstatSync(source).isDirectory()) {
+		return;
+	}
+
+	const files = fs.readdirSync(source);
+	files.forEach((file) => {
+		const curSource = path.join(source, file);
+		if (fs.lstatSync(curSource).isDirectory()) {
+			copyFolderRecursiveSync(curSource, targetFolder);
+			return;
+		}
+
+		copyFileSync(curSource, targetFolder);
 	});
 }
